checkpoint of edits in refactor
