+++ CHANGELOG.md
@@ -20,6 +20,7 @@ There is one key exception to the rules above -- and that is with `MAJOR`=0 rele
 - add `simmate database load-remote-archives` command and `load_remote_archives` utility that populates all tables from `database.third_parties`
 - add `load_default_sqlite3_build` utility that downloads a pre-built database with all third-party data present. This is an alternative to calling `load_all_remote_archives` if you are using sqlite3 and saves a significant amount of time for users.
 - standardize workflow naming. Note this breaks from python naming conventions for classes ([#150](https://github.com/jacksund/simmate/issues/150))
+- dynamically determine `register_kwargs` and rename property to `parameters_to_register`
 
 **Refactors**
 - move contents of `configuration.django.database` to `database.utilities`
