+++ src/simmate/calculators/vasp/workflows/diffusion/neb_from_images.py
@@ -1,17 +1,19 @@
 # -*- coding: utf-8 -*-
 
-from simmate.workflow_engine.workflow import Workflow
+from simmate.workflow_engine import Workflow, task
 from simmate.workflow_engine.common_tasks import load_input_and_register
 from simmate.calculators.vasp.tasks.nudged_elastic_band import (
     MITNudgedElasticBand,
 )
 from simmate.calculators.vasp.database.nudged_elastic_band import (
-    MITDiffusionAnalysis,
+    # MITDiffusionAnalysis,
     MITMigrationHop,
     MITMigrationImage,
 )
-from simmate.calculators.vasp.workflows.nudged_elastic_band.utilities import (
-    SaveNEBOutputTask,
+from simmate.database.base_data_types.nudged_elastic_band import (
+    # DiffusionAnalysis as DiffusionAnalysisTable,
+    MigrationHop as MigrationHopTable,
+    MigrationImage as MigrationImageTable,
 )
 
 
@@ -58,15 +60,49 @@ class Diffusion__Vasp__NebFromImages(Workflow):
 
         result = cls.s3task.run(**parameters_cleaned).result()
 
-        save_results_task = SaveNEBOutputTask(
-            MITDiffusionAnalysis,
-            MITMigrationHop,
-            MITMigrationImage,
-        )
-        calculation_id = save_results_task(
+        calculation_id = save_neb_results(
             output=result,
+            # migration_hop_table=MITDiffusionAnalysis,
+            migration_hop_table=MITMigrationHop,
+            migration_image_table=MITMigrationImage,
             diffusion_analysis_id=diffusion_analysis_id,
             migration_hop_id=migration_hop_id,
         )  # calculation_id will correspond to the migration hop table
 
         return result
+
+
+@task
+def save_neb_results(
+    output,
+    # diffusion_analyis_table: DiffusionAnalysisTable,
+    migration_hop_table: MigrationHopTable,
+    migration_image_table: MigrationImageTable,
+    diffusion_analysis_id: int = None,
+    migration_hop_id: int = None,
+):
+
+    # split our results and corrections (which are given as a dict) into
+    # separate variables
+    # Our result here is not a VaspRun object, but instead a NEBAnalysis
+    # object. See NudgedElasticBandTask.workup()
+    result = output["result"]
+
+    # TODO: These aren't saved for now. Consider making MigrationHopTable
+    # a Calculation and attaching these there.
+    corrections = output["corrections"]
+    directory = output["directory"]
+
+    # First, we need a migration_hop database object.
+    # All of hops should link to a diffusion_analysis entry, so we check
+    # for that here too. The key thing of these statements is that we
+    # have a migration_hop_id at the end.
+
+    migration_hop_db = migration_hop_table.from_pymatgen(
+        analysis=result,
+        diffusion_analysis_id=diffusion_analysis_id,
+        migration_hop_id=migration_hop_id,
+    )
+
+    # If the user wants to access results, they can do so through the hop id
+    return migration_hop_db.id
