+++ src/simmate/calculators/vasp/workflows/relaxation/quality02.py
@@ -12,7 +12,6 @@ from simmate.calculators.vasp.database.relaxation import (
 class Relaxation__Vasp__Quality02(Workflow):
     s3task = Quality02RelaxationTask
     database_table = Quality02RelaxationResults
-    register_kwargs = ["structure", "source"]
     description_doc_short = (
         "low-quality settings meant for highly unreasonable structures"
     )
