+++ src/simmate/calculators/vasp/workflows/static_energy/quality04.py
@@ -12,7 +12,6 @@ from simmate.calculators.vasp.database.energy import (
 class StaticEnergy__Vasp__Quality04(Workflow):
     s3task = Quality04EnergyTask
     database_table = Quality04StaticEnergyResults
-    register_kwargs = ["structure", "source"]
     description_doc_short = (
         "low-quality settings meant for highly unreasonable structures"
     )
